+++ app/components/captcha/component.rb
@@ -3,7 +3,7 @@ module Captcha
     attr_reader :show_checkbox_recaptcha, :captcha_required, :action
 
     def initialize(captcha_required:, show_checkbox_recaptcha:, action:)
-      super
+      super()
 
       @show_checkbox_recaptcha = show_checkbox_recaptcha
       @captcha_required = captcha_required