+++ app/components/common/auction_type_icon/component.rb
@@ -6,7 +6,7 @@ module Common
       def initialize(auction:)
         @auction = auction
 
-        super
+        super()
       end
 
       def english?