+++ app/components/common/form/telephone_field/component.rb
@@ -5,7 +5,7 @@ module Common
         attr_reader :attribute, :form, :options
 
         def initialize(form:, attribute:, options: {})
-          super
+          super()
 
           @form = form
           @attribute = attribute