+++ app/components/common/header/component.rb
@@ -6,7 +6,7 @@ module Common
       attr_reader :notifications
 
       def initialize(notifications:)
-        super
+        super()
 
         @notifications = notifications
       end