+++ app/components/common/static_notice/component.rb
@@ -4,7 +4,7 @@ module Common
       attr_accessor :current_user
 
       def initialize(current_user:)
-        super
+        super()
 
         @current_user = current_user
       end