+++ app/components/modals/pay_invoice/component.rb
@@ -4,7 +4,7 @@ module Modals
       attr_reader :invoice
 
       def initialize(invoice:)
-        super
+        super()
 
         @invoice = invoice
       end