+++ app/components/pages/auction/notification_subscribe/component.rb
@@ -5,7 +5,7 @@ module Pages
         attr_reader :current_user
 
         def initialize(current_user:)
-          super
+          super()
 
           @current_user = current_user
         end