+++ app/components/pages/offers/blind_auction_offer_table/component.rb
@@ -5,7 +5,7 @@ module Pages
         attr_reader :offer
 
         def initialize(offer:)
-          super
+          super()
 
           @offer = offer
         end
@@ -13,8 +13,7 @@ module Pages
         def header_collection
           [{ column: nil, caption: t('auctions.domain_name'), options: { class: '' } },
            { column: nil, caption: t('auctions.ends_at'), options: { class: '' } },
-           { column: nil, caption: t('offers.price'), options: { class: '' } },
-           { column: nil, caption: t('offers.total'), options: { class: '' } }]
+           { column: nil, caption: t('offers.price'), options: { class: '' } }]
         end
       end
     end