+++ app/components/pages/offers/offers_table/component.rb
@@ -5,7 +5,7 @@ module Pages
         attr_reader :offers
 
         def initialize(offers:)
-          super
+          super()
 
           @offers = offers
         end
@@ -14,7 +14,6 @@ module Pages
           [{ column: 'auctions.domain_name', caption: t('auctions.domain_name'), options: { class: 'sorting' } },
            { column: 'auctions.platform', caption: t('auctions.type'), options: { class: 'sorting' } },
            { column: nil, caption: t('offers.show.your_last_offer'), options: { class: '' } },
-           { column: nil, caption: t('offers.total'), options: { class: '' } },
            { column: 'auctions.ends_at', caption: t('auctions.ends_at'),
              options: { class: 'sorting' } },
            { column: nil, caption: t('result_name'), options: {} },