+++ config/application.rb
@@ -11,6 +11,7 @@ module AuctionCenter
   class Application < Rails::Application
     # Initialize configuration defaults for originally generated Rails version.
     config.load_defaults 7.0
+    config.active_support.to_time_preserves_timezone = :zone
     # config.autoloader = :classic
     # config.autoloader = :zeitwerk